Separate translation into a module
...so the translation can be shared with other
modules. Opens the possiblity of spllitting the UI
code into multiple files.
