+++ carveracontroller/__main__.py
@@ -1,9 +1,5 @@
-from carveracontroller.main import main, init_lang, Lang
-
-# tr is used throughout the kivvy .kv definition files
-# thus kivy expects it be availiable from the caller
-default_lang = init_lang()
-tr = Lang(default_lang)
+from carveracontroller.main import main
+from carveracontroller.translation import tr
 
 if __name__ == "__main__":
     main()
